requirements.txt
PyQt6>=6.0.0
requests>=2.25.1
beautifulsoup4>=4.9.3
python-docx>=0.8.11
openpyxl>=3.0.7
python-pptx>=0.6.21
html2text>=2020.1.16
pypandoc>=1.5
Pillow>=8.2.0